Move user activity related coverage to concern spec
